M13 #171-173, 175-176, 178-180
The missing cards are Garruk, Primal Hunter (planeswalker, duh); and Mwonvuli Beast Tracker. The Tracker feels like it can be done, but it's inconvenient and I'm not sure how. The easiest route would seem to be recoding all cards so that their keywords are in effect in the library (would this cause other problems?), and mess with the filter so it doesn't require all four keywords (can you use the "return ..." type filter for choosing targets?).