Big restructuring (of my stuff), not fully tested
The DLC_1000 is intended to be storehouse for all of the universal stuff - sort of like Kev's DLC, but importantly not having a campaign so it won't cause any conflicts with whichever campaign expansion you want to use. I also don't see much reason to put cards in it, other than the mana tokens that can be used all over the place.